Improve converter error handling #389
