+++ ktorfit-lib-common/src/commonMain/kotlin/de/jensklingenberg/ktorfit/converter/builtin/DefaultSuspendResponseConverterFactory.kt
@@ -2,6 +2,7 @@ package de.jensklingenberg.ktorfit.converter.builtin
 
 import de.jensklingenberg.ktorfit.Ktorfit
 import de.jensklingenberg.ktorfit.converter.Converter
+import de.jensklingenberg.ktorfit.converter.KtorfitResult
 import de.jensklingenberg.ktorfit.internal.TypeData
 import io.ktor.client.statement.*
 
@@ -14,7 +15,20 @@ internal class DefaultSuspendResponseConverterFactory : Converter.Factory {
     class DefaultSuspendResponseConverter(val typeData: TypeData) :
         Converter.SuspendResponseConverter<HttpResponse, Any> {
         override suspend fun convert(response: HttpResponse): Any {
-            return response.call.body(typeData.typeInfo)
+            return convert(KtorfitResult.Success(response))
+        }
+
+        override suspend fun convert(result: KtorfitResult): Any {
+            return when (result) {
+                is KtorfitResult.Failed -> {
+                    throw result.throwable
+                }
+
+                is KtorfitResult.Success -> {
+                    val response = result.response
+                    response.call.body(typeData.typeInfo)
+                }
+            }
         }
     }
